backend/requirements.txt
aiofile==3.8.5
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.18.0
alembic==1.10.2
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
billiard==3.6.4.0
black==23.1.0
caio==0.9.12
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.26.0
coverage==7.2.2
exceptiongroup==1.1.1
fastapi==0.95.0
fastapi-pagination==0.12.1
ffmpeg-python==0.2.0
filelock==3.10.0
flake8==6.0.0
frozenlist==1.3.3
future==0.18.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
kombu==5.2.4
llvmlite==0.39.1
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
openai-whisper==20230314
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
pytest==7.2.2
pytest-asyncio==0.21.0
pytest-celery==0.0.0
pytest-cov==4.0.0
pytest_async==0.1.1
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.1
regex==2022.10.31
requests==2.28.2
simplejson==3.18.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
sqlalchemy2-stubs==0.0.2a33
starlette==0.26.1
sympy==1.11.1
tiktoken==0.3.1
tomli==2.0.1
torch==2.0.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
